requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
EasyProcess==0.2.3
funcsigs==1.0.2
idna==2.8
more-itertools==4.3.0
pathlib2==2.3.3
pluggy==0.8.0
py==1.7.0
PyMsgBox==1.0.6
PyScreeze==0.1.18
python-xlib==0.23
PyTweening==1.0.3
PyVirtualDisplay==0.2.1
requests==2.21.0
scandir==1.9.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
